app/requirements/req-base.txt
# LTS Django 3.2.x
Django==3.2.18

# Django packages
django-celery-beat
django-celery-email
django-celery-results
django-cors-headers
django-extensions
django-filter
django-rest-swagger
django-storages

# Database
psycopg2-binary

# Requests
requests

# REST
djangorestframework
djangorestframework-gis

# Datapunt packages
datapunt-authorization-django
drf_amsterdam
drf-extensions
flex

# Documentation
coreapi

# Celery
Celery

# Sentry
raven

#Azure insight
opencensus-ext-azure

# PDF
weasyprint==52.5
cairocffi
CairoSVG
lxml

# Images
Pillow

# Date util
python-dateutil

# Elasticsearch
elasticsearch-dsl==6.4.0

# dsl
Arpeggio
Click
textX

# Email templates
mistune

# Filetype scanning
python-magic

# Graph handling
networkx

# Markdown
django-markdownx
Markdown

uWSGI